LPS-87016 Separate odata conversion functions
