+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/utils/constants.es.js
@@ -1,0 +1,27 @@
+export const AND = 'and';
+
+export const BOOLEAN = 'boolean';
+
+export const CONTAINS = 'contains';
+
+export const DATE = 'data';
+
+export const EQ = 'eq';
+
+export const GE = 'ge';
+
+export const GROUP = 'group';
+
+export const GT = 'gt';
+
+export const LE = 'le';
+
+export const LT = 'lt';
+
+export const NE = 'ne';
+
+export const NUMBER = 'number';
+
+export const OR = 'or';
+
+export const STRING = 'string';
